[bug][upstream] Clear errors before save.
Otherwise errors from previous save attempts will remain, so errors could
not be empty for a successful save attempt, and errors could include
irrelevant or duplicate errors messages for an unsuccessful save.

Bug reported and patch submitted upstream, but haven't received a response
since October 22. This commit is compatible with the upstream patch.