Add support for building the package in Python 3.12
As discussed in the [issue](https://github.com/Shopify/shopify_python_api/issues/725)
there is an issue when trying to install this package when using
Python 3.12. The issue is caused by
[PEP-632](https://peps.python.org/pep-0632/) removing `distutils` from
the built-in libraries this version of python is shipped with. In
versions 6.0.0 and lower of PyYaml, this causes the build the fail. As
mentioned in their [issue](https://github.com/yaml/pyyaml/issues/756)
using v6.0.1 or later will resolve this issue. So updating our
`setup.py` install_requires dependency specifications to reflect this.

While I'm here, also update the classifiers to include support for
PY3.12 and show this more clearly on the README of the repo.